requirements.txt
agno[os]
aioboto3
llama-cloud
openpyxl
python-dotenv
pydantic
python-docx
fastapi[standard]
